add pinview to anything that uses PlayerTableViewController
